Use arista.go
Move platform detection to internal/helpers/arista.go

Cleanup
